fix(assets-ext): align decimals to BIGINT/Long, drop dead lazy hints
Three small entity/schema hygiene fixes from the JPA-vs-SQL audit:

1. decimals widened from INTEGER to BIGINT in cip26_metadata and
   cip68_metadata migrations (postgres/h2/mysql). The Java pipeline
   (parser, ParsedCip68Datum, FungibleTokenMetadata, LongProperty,
   entity field) was already Long; the column was the lone INTEGER
   outlier, forcing boundary conversions. Spec range is 0-19 either
   way, so the 4-byte-per-row overhead is negligible (~50 KB at 25k
   rows), and the pipeline is now uniformly Long.

2. Cip68Metadata.slot gets an explicit @Column(nullable=false). JPA
   defaults @Id to NOT NULL anyway, but the other two PK fields
   (policy_id, asset_name) had it explicitly — adding here matches.

3. Removed @Basic(fetch=LAZY) from logo and datum on Cip68Metadata.
   Lazy basic-field fetch requires Hibernate bytecode enhancement
   (org.hibernate.orm gradle plugin), which this build does not
   configure — the annotations were silently ignored, with both
   columns loading eagerly anyway. Comments now point at the gradle
   plugin if anyone wants to revive proper lazy fetch (datum is the
   highest-impact target — typically 1-10 KB per row).

Note: existing deployments have the V0_1600_1 migration already
applied with INTEGER. Picking up the BIGINT change requires either
manual ALTER TABLE + flyway repair, or a fresh re-sync. Pre-merge,
modifying the initial migration in place is acceptable.

+++ extensions/assets-ext/src/main/resources/db/store/postgresql/V0_1600_1__init.sql
@@ -37,8 +37,10 @@ CREATE TABLE cip26_metadata (
     -- see MetadataValidationRules.MAX_DESCRIPTION_LENGTH in cf-tokens-cip26).
     description    VARCHAR(500),
     -- CIP-26 decimals: spec range [0, 19] inclusive (well-known property 'decimals').
-    -- INTEGER is oversized but standard; SMALLINT would also work.
-    decimals       INTEGER,
+    -- BIGINT to align with the rest of the pipeline (Java Long throughout: parser,
+    -- entity, DTO). INTEGER would suffice for spec values but introduces type-conversion
+    -- noise at every boundary; the 4-byte overhead per row is negligible.
+    decimals       BIGINT,
     -- base64-encoded image, up to ~87400 chars per CIP-26 — kept as TEXT.
     -- Postgres TOAST handles the size transparently; merging the column here keeps
     -- one row per subject and removes the need for a separate ft_offchain_logo table.
@@ -85,7 +87,9 @@ CREATE TABLE cip68_metadata (
     -- CIP-68 'url': aligns with CIP-26 url cap of 250 chars.
     url            VARCHAR(250),
     -- CIP-68 'decimals': unsigned integer in the datum. In practice 0–19.
-    decimals       INTEGER,
+    -- BIGINT to match the rest of the pipeline (Java Long); see cip26_metadata.decimals
+    -- comment above for the rationale.
+    decimals       BIGINT,
     -- base64-encoded logo (PNG/JPG/SVG). Variable-length, can be tens of KB.
     logo           TEXT,
     -- CIP-68 NFT 'image': IPFS / data: URL / https URL.
